# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bug Reproduction Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tools/env-diff:
    post:
      operationId: envDiff
      tags: [tools]
      summary: Compare two environment configs and identify bug-causing differences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnvDiffBody"
      responses:
        "200":
          description: Environment diff result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvDiffResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /tools/nl2test:
    post:
      operationId: nl2test
      tags: [tools]
      summary: Generate a test case from a natural language description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Nl2TestBody"
      responses:
        "200":
          description: Generated test code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Nl2TestResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /tools/flaky-detector:
    post:
      operationId: flakyDetector
      tags: [tools]
      summary: Detect flaky tests and explain why they are flaky
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlakyDetectorBody"
      responses:
        "200":
          description: Flakiness analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlakyDetectorResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
...
    EnvDiffBody:
      type: object
      properties:
        env1:
          type: string
          description: First environment config (key=value lines or JSON)
        env2:
          type: string
          description: Second environment config (key=value lines or JSON)
        bugDescription:
          type: string
          description: Description of the bug or intermittent behavior
        label1:
          type: string
          description: Label for first environment (e.g. "Local", "Staging")
        label2:
          type: string
          description: Label for second environment (e.g. "Production", "CI")
      required:
        - env1
        - env2
        - bugDescription
    EnvDiffResult:
      type: object
      properties:
        differences:
          type: array
          items:
            $ref: "#/components/schemas/EnvDifference"
        verdict:
          type: string
          description: AI verdict on the most likely culprit
        likelihood:
          type: string
          enum: [high, medium, low]
        summary:
          type: string
      required:
        - differences
        - verdict
        - likelihood
        - summary
    EnvDifference:
      type: object
      properties:
        key:
          type: string
        value1:
          type: ["string", "null"]
        value2:
          type: ["string", "null"]
        impact:
          type: string
          enum: [critical, likely, unlikely, irrelevant]
        reasoning:
          type: string
      required:
        - key
        - value1
        - value2
        - impact
        - reasoning
    Nl2TestBody:
      type: object
      properties:
        description:
          type: string
          description: Natural language description of what to test
        framework:
          type: string
          description: Test framework hint (e.g. Jest, Pytest, Mocha, Cypress)
        codeContext:
          type: string
          description: Optional relevant code context
      required:
        - description
    Nl2TestResult:
      type: object
      properties:
        testCode:
          type: string
        framework:
          type: string
        explanation:
          type: string
        coverageNotes:
          type: string
      required:
        - testCode
        - framework
        - explanation
        - coverageNotes
    FlakyDetectorBody:
      type: object
      properties:
        testCode:
          type: string
          description: The test suite or individual test code to analyze
        language:
          type: string
          description: Language/framework hint
        runCount:
          type: integer
          description: Optional number of times the test has been run
      required:
        - testCode
    FlakyDetectorResult:
      type: object
      properties:
        flakyTests:
          type: array
          items:
            $ref: "#/components/schemas/FlakyTest"
        overallRisk:
          type: string
          enum: [high, medium, low, none]
        summary:
          type: string
      required:
        - flakyTests
        - overallRisk
        - summary
    FlakyTest:
      type: object
      properties:
        testName:
          type: string
        riskLevel:
          type: string
          enum: [high, medium, low]
        category:
          type: string
          enum: [race_condition, environment_dependency, non_deterministic_data, timing, external_dependency, state_leak, other]
        explanation:
          type: string
        fix:
          type: string
      required:
        - testName
        - riskLevel
        - category
        - explanation
        - fix
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error